libct/userns: split userns detection from internal userns code
Commit 4316df8b53cce2933d7f0f3cf4bf87da67459835 isolated RunningInUserNS
to a separate package to make it easier to consume without bringing in
additional dependencies, and with the potential to move it separate in
a similar fashion as libcontainer/user was moved to a separate module
in commit ca32014adbc5bcdd1ec5862672c35c7d4cbd849d. While RunningInUserNS
is fairly trivial to implement, it (or variants of this utility) is used
in many codebases, and moving to a separate module could consolidate
those implementations, as well as making it easier to consume without
large dependency trees (when being a package as part of a larger code
base).

Commit 1912d5988bbb379189ea9ceb2e03945738c513dc and follow-ups introduced
cgo code into the userns package, and code introduced in those commits
are not intended for external use, therefore complicating the potential
of moving the userns package separate.

This commit moves the new code to a separate package; some of this code
was included in v1.1.11 and up, but I could not find external consumers
of `GetUserNamespaceMappings` and `IsSameMapping`. The `Mapping` and
`Handles` types (added in ba0b5e26989f39d0bdadeeff38182902df781df6) only
exist in main and in non-stable releases (v1.2.0-rc.x), so don't need
an alias / deprecation.
