+++ libcontainer/internal/userns/usernsfd_linux.go
